+++ apps/web/app/setup/organization/[organizationId]/invite/types/invites.ts
@@ -1,0 +1,25 @@
+import { z } from "zod";
+import { ZInvite } from "@formbricks/database/zod/invites";
+import { ZUserName } from "@formbricks/types/user";
+
+export const ZInvitee = ZInvite.pick({
+  email: true,
+  role: true,
+  teamIds: true,
+}).extend({
+  name: ZUserName,
+});
+
+export type TInvitee = z.infer<typeof ZInvitee>;
+
+export const ZInviteMembersFormSchema = z.record(
+  ZInvite.pick({
+    email: true,
+    name: true,
+  }).extend({
+    email: z.string().email("Invalid email address"),
+    name: ZUserName,
+  })
+);
+
+export type TInviteMembersFormSchema = z.infer<typeof ZInviteMembersFormSchema>;
