+++ apps/web/lib/utils/services.ts
@@ -1,6 +1,7 @@
 "use server";
 
 import { contactCache } from "@/lib/cache/contact";
+import { inviteCache } from "@/lib/cache/invite";
 import { teamCache } from "@/lib/cache/team";
 import { Prisma } from "@prisma/client";
 import { cache as reactCache } from "react";
@@ -11,7 +12,6 @@ import { cache } from "@formbricks/lib/cache";
 import { segmentCache } from "@formbricks/lib/cache/segment";
 import { environmentCache } from "@formbricks/lib/environment/cache";
 import { integrationCache } from "@formbricks/lib/integration/cache";
-import { inviteCache } from "@formbricks/lib/invite/cache";
 import { projectCache } from "@formbricks/lib/project/cache";
 import { responseCache } from "@formbricks/lib/response/cache";
 import { responseNoteCache } from "@formbricks/lib/responseNote/cache";
