+++ apps/web/modules/auth/signup/types/invites.ts
@@ -1,0 +1,8 @@
+import { Invite } from "@prisma/client";
+
+export interface InviteWithCreator extends Invite {
+  creator: {
+    name: string | null;
+    email: string;
+  };
+}
