+++ apps/web/modules/ee/role-management/types/invites.ts
@@ -1,0 +1,8 @@
+import { z } from "zod";
+import { ZInvite } from "@formbricks/database/zod/invites";
+
+export const ZInviteUpdateInput = ZInvite.pick({
+  role: true,
+});
+
+export type TInviteUpdateInput = z.infer<typeof ZInviteUpdateInput>;
