+++ apps/web/modules/organization/settings/teams/components/edit-memberships/members-info.tsx
@@ -1,12 +1,12 @@
-import { isInviteExpired } from "@/app/lib/utils";
 import { EditMembershipRole } from "@/modules/ee/role-management/components/edit-membership-role";
 import { MemberActions } from "@/modules/organization/settings/teams/components/edit-memberships/member-actions";
+import { isInviteExpired } from "@/modules/organization/settings/teams/lib/utilts";
+import { TInvite } from "@/modules/organization/settings/teams/types/invites";
 import { Badge } from "@/modules/ui/components/badge";
 import { TooltipRenderer } from "@/modules/ui/components/tooltip";
 import { useTranslations } from "next-intl";
 import { getAccessFlags } from "@formbricks/lib/membership/utils";
 import { getFormattedDateTimeString } from "@formbricks/lib/utils/datetime";
-import { TInvite } from "@formbricks/types/invites";
 import { TMember, TOrganizationRole } from "@formbricks/types/memberships";
 import { TOrganization } from "@formbricks/types/organizations";
