+++ apps/web/modules/organization/settings/teams/lib/utilts.ts
@@ -1,4 +1,4 @@
-import { TInvite } from "@formbricks/types/invites";
+import { TInvite } from "@/modules/organization/settings/teams/types/invites";
 
 export const isInviteExpired = (invite: TInvite) => {
   const now = new Date();
